Scale the pixel data instead of the limit
When calculating the warm pixels, scale the pixel data (img) instead
of the limit (using get_limit) and count the pixels with values > 100.

It should be mathematically equivalent, but it is less confusing.

Also use a finer range for the "possible temps" for the lookup table
and put it out to a "labeled" npsave file as well as a pickle.